refactor(sdk): extract shared aggregate-count helper; split SplitCounts into own file
`DocumentSplitCounts::FromProof` used to delegate its `group_by =
[]` branch to `DocumentCount::FromProof` via cross-impl call. That
worked but read as architectural awkwardness — readers don't
expect one trait impl to invoke another's `maybe_from_proof_with_metadata`
to do its job, and it left the per-shape proof verification body
buried inside `DocumentCount`.

Pull the per-shape dispatch into a free function
`verify_aggregate_count(request, response, version, provider)` in
a new `count_proof_helpers.rs` sibling module. Both impls become
thin wrappers around it:
- `DocumentCount::FromProof` calls it once, maps `Option<u64>` to
  `Option<DocumentCount>`.
- `DocumentSplitCounts::FromProof` (aggregate branch) calls it,
  maps `Option<u64>` to a single empty-key entry; per-group
  branches keep their existing direct-verifier dispatch since
  they emit shaped entries from the verifier, not a sum.

Same opportunity to fix the file sizes: `document_count.rs` was
~600 LoC carrying two `FromProof` impls. Split into
`document_count.rs` (~55 LoC) and `document_split_counts.rs`
(~225 LoC). Shared helpers (`assert_select_is_count`,
`limit_to_u16_or_default`, `verify_aggregate_count`) live in
`count_proof_helpers.rs` and are `pub(super)`.

No behavior change. All 7 SDK fetch tests still pass.
