+++ packages/rs-sdk/src/platform/documents/mod.rs
@@ -1,3 +1,5 @@
+pub(super) mod count_proof_helpers;
 pub mod document_count;
 pub mod document_query;
+pub mod document_split_counts;
 pub mod transitions;
